Split Angular integration test coverage
Move the real-editor CKEditor component checks into a dedicated integration spec and define an addressable Vitest integration subset so CI can target version-sensitive tests separately from the full suite.
